# src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UserCategoryName": {
				"enum": [
					"Aluno",
					"Professor",
					"Bibliotecário"
				],
				"type": "string"
			},
			"CourseName": {
				"enum": [
					"ADS",
					"Pedagogia",
					"Administração"
				],
				"type": "string"
			},
			"CreateUserDto": {
				"properties": {
					"nome": {
						"type": "string"
					},
					"cpf": {
						"type": "string"
					},
					"categoria": {
						"$ref": "#/components/schemas/UserCategoryName"
					},
					"curso": {
						"$ref": "#/components/schemas/CourseName"
					}
				},
				"required": [
					"nome",
					"cpf",
					"categoria",
					"curso"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BasicResponseDto": {
				"properties": {
					"message": {
						"type": "string"
					},
					"object": {}
				},
				"required": [
					"message",
					"object"
				],
				"type": "object",
				"additionalProperties": false
			},
			"userActive": {
				"enum": [
					"ativo",
					"inativo",
					"suspenso"
				],
				"type": "string"
			},
			"UpdateUserDto": {
				"properties": {
					"nome": {
						"type": "string"
					},
					"categoria": {
						"$ref": "#/components/schemas/UserCategoryName"
					},
					"curso": {
						"$ref": "#/components/schemas/CourseName"
					},
					"status": {
						"$ref": "#/components/schemas/userActive"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"CreateStockDto": {
				"properties": {
					"codigo_exemplar": {
						"type": "number",
						"format": "double"
					},
					"isbn": {
						"type": "string"
					}
				},
				"required": [
					"codigo_exemplar",
					"isbn"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateLoanDto": {
				"properties": {
					"codigo_exemplar": {
						"type": "number",
						"format": "double"
					},
					"cpf": {
						"type": "string"
					}
				},
				"required": [
					"codigo_exemplar",
					"cpf"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateBookDto": {
				"properties": {
					"titulo": {
						"type": "string"
					},
					"autor": {
						"type": "string"
					},
					"editora": {
						"type": "string"
					},
					"edicao": {
						"type": "string"
					},
					"ISBN": {
						"type": "string"
					},
					"categoria": {
						"type": "string"
					}
				},
				"required": [
					"titulo",
					"autor",
					"editora",
					"edicao",
					"ISBN",
					"categoria"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BookCategoryEnum": {
				"enum": [
					"Romance",
					"Computação",
					"Letras",
					"Gestão"
				],
				"type": "string"
			},
			"UpdateBookDto": {
				"properties": {
					"titulo": {
						"type": "string"
					},
					"autor": {
						"type": "string"
					},
					"editora": {
						"type": "string"
					},
					"edicao": {
						"type": "string"
					},
					"categoria": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "projeto-1",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/usuarios": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"usuarios"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserDto"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "ListUsers",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"usuarios"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "nome",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "cpf",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "categoria",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/UserCategoryName"
						}
					},
					{
						"in": "query",
						"name": "curso",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/CourseName"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/userActive"
						}
					}
				]
			}
		},
		"/usuarios/{cpf}": {
			"get": {
				"operationId": "FindUserByCpf",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"usuarios"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "cpf",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUserByCpf",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"usuarios"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "cpf",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUserByCpf",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"usuarios"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "cpf",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/estoque": {
			"post": {
				"operationId": "CreateCopy",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"estoque"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateStockDto"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "ListCopies",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"estoque"
				],
				"security": [],
				"parameters": []
			}
		},
		"/estoque/{codigo}": {
			"get": {
				"operationId": "FindCopyById",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"estoque"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "codigo",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateAvailabilityById",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"estoque"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "codigo",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"disponivel": {
										"type": "boolean"
									}
								},
								"required": [
									"disponivel"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCopyById",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"estoque"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "codigo",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/emprestimos": {
			"post": {
				"operationId": "CreateLoan",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"emprestimos"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateLoanDto"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "ListLoans",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"emprestimos"
				],
				"security": [],
				"parameters": []
			}
		},
		"/emprestimos/{id}/devolucao": {
			"put": {
				"operationId": "UpdateReturnDateById",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"emprestimos"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/livros": {
			"post": {
				"operationId": "CreateBook",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"livros"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateBookDto"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "ListBooks",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"livros"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "titulo",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "isbn",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "autor",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "editora",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "edicao",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "categoria",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/BookCategoryEnum"
						}
					}
				]
			}
		},
		"/livros/{isbn}": {
			"get": {
				"operationId": "FindBookByIsbn",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"livros"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "isbn",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateBookByIsbn",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"livros"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "isbn",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateBookDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteBookByIsbn",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"livros"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "isbn",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}